Adds Docker
What
---

I added a docker-compose file that spins up a minio and postgres
container for this app. In prod I won't use either, but I wanted to use
them locally so I could work out kinks for using object storage and a
postgres db.

Why
---

In prod I'm going to host on Digital Ocean and use their object storage
and a managed DB. I could go lighter and put it all on a single VPS but
the point of this is to figure out how I would deploy a real-world app.